chore: shared seeder utils
